add ThrowableCallable to "remove code dupe"
I'm not sure if I'm 100% sold on this. Main goals are
- use `Instrumentor` in `InstrumentingInterceptor`
- move package-scoped `Instrumentation` helpers back onto
 `Instrumentor` as private methods

Changes
----------

I've added a `ThrowableCallable<T>`, basically a `Callable`
that throws `Throwable` instead of `Exception`.

```java
public interface ThrowableCallable<T> {
    T call() throws Throwable;
}
```

Added `instrumenting`, `callThrowably` methods to `Instrumentor`.
(any better name suggestions?)

We need this if we want to use `Instrumentor` in the interceptor, because
`MethodInvocation#proceed` throws `Throwable`. Starting to feel a little
yak shave-y, but enumerating every type combination might just be
something you do in a library like this.

This means another big method that returns a lambda, so we're not
actually de-duping a *ton* of code here, which is another reason
I'm dubious about the value of this change.

One question: in `InstrumentingInterceptor`, we already have an
`Optional<Double>`, so I've made the `Instrumentor#callThrowably` that takes an
`Optional<Double>` package-scoped. That said, it's an argument for making all the
methods that take `Optional<Double>` for the error threshold into public
methods.

Next step: now if we want we can move all the package-scoped stuff in
`Instrumentation` back onto private methods on the Instrumentor, which I
think is another big win in terms of organization and clarity.