Support for instrumentation of classes
`InstrumentedAnnotations` now binds an `InstrumentingInterceptor` to
methods of a class annotated with `@Instrumented`.

Getting the `@Instrumented` for a single annotated method is different
than getting it for all the public methods of a class, so:

- Refactor means of getting instance of `Instrumented` from
  `MethodInvocation` into `InstrumentationDetails` interface
- Refactor means of naming a single method into same
- Implemented by `MethodInstrumentation` (old behavior) and
  `ClassInstrumentation` (new behavior)
- I considered adding unit tests for this, but since we cant mock
  `Method` or `Class`, there's not much else we can test outside
  of what is already tested in `InstrumentingInterceptorTest`

To make use of the class-level interception we bind an additional
interceptor,
at the same time allowing `@Instrumented` attributes on a method
to take precedence over `@Instrumented` on a class.

Also
----

- Clean up instances of `getMeters` in readme that should have been
  `getTimers`
- Move lambda `any -> true` in `Builder` inner classes to
  `ExceptionFilters#markAllExceptions`